+++ Sandbox/GameImpl.cs
@@ -365,6 +365,10 @@ namespace Sandbox {
                 Console.WriteLine("Down: " + string.Join(", ", Input.InputsDown));*/
             if (MlemGame.Input.InputsPressed.Length > 0)
                 Console.WriteLine("Pressed: " + string.Join(", ", MlemGame.Input.InputsPressed));
+            MlemGame.Input.HandleKeyboardRepeats = false;
+            Console.WriteLine("Down time: " + MlemGame.Input.GetDownTime(Keys.A));
+            Console.WriteLine("Time since press: " + MlemGame.Input.GetTimeSincePress(Keys.A));
+            Console.WriteLine("Up time: " + MlemGame.Input.GetUpTime(Keys.A));
         }
 
         protected override void DoDraw(GameTime gameTime) {